+++ MaxMind.MinFraud.UnitTest/Response/IPAddressTest.cs
@@ -52,9 +52,6 @@ namespace MaxMind.MinFraud.UnitTest.Response
                 address.RiskReasons[1].Reason
             );
             Assert.Equal(time, address.Location.LocalTime?.ToString("yyyy-MM-ddTHH:mm:ssK"));
-#pragma warning disable CS0618 // Type or member is obsolete
-            Assert.True(address.Country.IsHighRisk);
-#pragma warning restore CS0618 // Type or member is obsolete
             Assert.True(address.Traits.IsAnonymous);
             Assert.True(address.Traits.IsAnonymousVpn);
             Assert.True(address.Traits.IsHostingProvider);