+++ MaxMind.MinFraud/Request/CreditCard.cs
@@ -76,41 +76,6 @@ namespace MaxMind.MinFraud.Request
         }
 
         /// <summary>
-        /// Legacy constructor for backwards compatibility.
-        /// </summary>
-        [Obsolete("Legacy constructor for backwards compatibility")]
-        public CreditCard(
-            string? issuerIdNumber,
-            string? last4Digits,
-            string? bankName,
-            string? bankPhoneCountryCode,
-            string? bankPhoneNumber,
-            char? avsResult,
-            char? cvvResult,
-            string? token,
-            bool? was3DSecureSuccessful
-        ) : this(issuerIdNumber, last4Digits, bankName, bankPhoneCountryCode, bankPhoneNumber, avsResult, cvvResult, token, was3DSecureSuccessful, null)
-        {
-        }
-
-        /// <summary>
-        /// Legacy constructor for backwards compatibility.
-        /// </summary>
-        [Obsolete("Legacy constructor for backwards compatibility")]
-        public CreditCard(
-            string? issuerIdNumber,
-            string? last4Digits,
-            string? bankName,
-            string? bankPhoneCountryCode,
-            string? bankPhoneNumber,
-            char? avsResult,
-            char? cvvResult,
-            string? token
-        ) : this(issuerIdNumber, last4Digits, bankName, bankPhoneCountryCode, bankPhoneNumber, avsResult, cvvResult, token, null)
-        {
-        }
-
-        /// <summary>
         /// The issuer ID number for the credit card. This is the first 6
         /// digits of the credit card number. It identifies the issuing bank.
         /// </summary>